+++ NEWS.md
@@ -1,5 +1,7 @@
 # odbc (development version)
 
+* SQL Server: Support for `DATETIMEOFFSET` data type (#918). 
+
 * Fix hang when parsing exceptionally long database errors (#916).
 
 * DB2: Fix error when writing to temp tables (#910).